requirements.txt
dirsync==2.2.5
imageio==2.31.1
moviepy==1.0.3
mutagen==1.46.0
natsort==8.4.0
numpy==1.25.2
numpy==1.21.5
opencv_python==4.8.0.76
Pillow==10.0.0
Pillow==9.0.1
Pillow==10.0.0
PyAutoGUI==0.9.54
pydub==0.25.1
pyscreenshot==3.1
